test(postcontroller): Adiciona testes e unifica testes de GET e o restante do CRUD
